Do not use github.com/simplereach/timeutils for windows
