+++ .gitignore
@@ -1,2 +1,3 @@
 /zabbixctl
+/zabbixctl.exe
 /zabbixctl.test
